drf/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.1.3
djangorestframework==3.14.0
Pillow==9.3.0
pip==22.3.1
PyMySQL==1.0.2
pytz==2022.6
setuptools==65.5.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.6
